Create SearchInRotatedArray.cpp
A sorted array 'arr' consisting of 'n' elements with an integer 'k'. The array is rotated and the problem is to find index at which 'k' is present in 'arr'. 

I've solved this problem using Binary Search approach. Initially , find out the pivot element and index of rotated array and then apply Binary Search approach to find the index.